@@ -4,43 +4,48 @@
   "main": "app/main.js",
   "author": "Simeon Ackermann, Konrad Abicht",
   "license": "GPL-3.0",
+  "version": "1.0.0",
+  "keywords": [],
+  "directories": {
+    "test": "test"
+  },
   "scripts": {
-    "lint": "eslint src/",
-    "start": "gulp",
-    "test": "jest",
-    "test:watch": "npm test -- --watch",
-    "build-dist": "webpack -p --progress --config ./webpack.production.config.js && echo '\n\n========================================\n\nDo not forget to push the dist-changes!\n\n========================================'"
-  },
-  "dependencies": {
-    "js-cookie": "^2.1.3",
-    "leaflet": "^0.7.7",
-    "promise-polyfill": "^6.0.2",
-    "react": "^15.0.0",
-    "react-bootstrap": "^0.30.1",
-    "react-dom": "^15.0.0",
-    "react-leaflet": "^0.12.3",
-    "react-swipeable": "^3.7.0",
-    "react-tap-event-plugin": "^2.0.1",
-    "whatwg-fetch": "^2.0.1"
-  },
-  "devDependencies": {
-    "babel-core": "^6.9.0",
-    "babel-jest": "^18.0.0",
-    "babel-loader": "^6.2.4",
-    "babel-polyfill": "^6.23.0",
-    "babel-preset-es2015": "^6.9.0",
-    "babel-preset-react": "^6.5.0",
-    "enzyme": "^2.7.1",
-    "jest": "^18.1.0",
-    "react-addons-test-utils": "^15.4.2",
-    "sinon": "^2.1.0",
-    "webpack": "^1.13.1",
-    "webpack-dev-server": "^1.14.1"
+    "test": "./node_modules/jest/bin/jest.js --coverage --coverageDirectory ./coverage",
+    "start": "webpack-dev-server --config ./webpack.config.js --mode development",
+    "build": "webpack --config ./webpack.production.config.js --mode production"
   },
   "babel": {
     "presets": [
-      "react",
-      "es2015"
+      "env",
+      "react"
     ]
+  },
+  "devDependencies": {
+    "babel-core": "^6.26.0",
+    "babel-loader": "^7.1.4",
+    "babel-preset-env": "^1.6.1",
+    "babel-preset-react": "^6.24.1",
+    "enzyme": "^3.3.0",
+    "enzyme-adapter-react-16": "^1.1.1",
+    "jest": "^22.4.3",
+    "jsdom": "^11.10.0",
+    "sinon": "^5.0.6",
+    "webpack": "^4.5.0",
+    "webpack-cli": "^2.0.14",
+    "webpack-dev-server": "^3.1.1"
+  },
+  "dependencies": {
+    "axios": "^0.18.0",
+    "js-cookie": "^2.2.0",
+    "leaflet": "^1.3.1",
+    "promise-polyfill": "^7.1.2",
+    "prop-types": "^15.6.1",
+    "react": "^16.3.1",
+    "react-bootstrap": "^0.32.1",
+    "react-dom": "^16.3.1",
+    "react-leaflet": "^1.9.1",
+    "react-prop-types": "^0.4.0",
+    "react-swipeable": "^4.2.1",
+    "react-tap-event-plugin": "^3.0.2"
   }
 }
