@@ -9,6 +9,10 @@ import A11yIcon from '../A11yIcon';
 
 import {getElement} from '../../utils/GuiUtils'
 
+let handleMouseOutTimeout = null;
+let handleMouseOverTimeout = null;
+let mouseOverBuilding = null;
+
 /**
  * Results entry component, renders a result with title, a11y icons and details
  */
@@ -19,14 +23,23 @@ class Entry extends React.Component {
         this.state = {
             stores: props.stores,
             building: props.building,
+            a11yIcons: new A11yIcon({building: props.building}),
+            a11yIconShowDetails: {},
             isLoading: false
         };
+
+        // dont show a11yicon details on start
+        this.state.a11yIcons.getAll().map((entry, id) => {
+            this.state.a11yIconShowDetails[entry] = false;
+        });
 
         this.handleShowDetails = this.handleShowDetails.bind(this);
         this.handleHideDetails = this.handleHideDetails.bind(this);
         this.handleMouseOver = this.handleMouseOver.bind(this);
+        this.handleMouseOut = this.handleMouseOut.bind(this);
         this.handleShowOnMap = this.handleShowOnMap.bind(this);
         this.handleMayHideSidebar = this.handleMayHideSidebar.bind(this);
+        this.a11yIconToggleShowDetails = this.a11yIconToggleShowDetails.bind(this);
     }
 
     componentWillReceiveProps(nextProps) {
@@ -88,8 +101,51 @@ class Entry extends React.Component {
         e.preventDefault();
     }
 
-    handleMouseOver(e) {
-        // @todo may show current hovered entry on map
+    /**
+     * Handles mouse hovers entry
+     * @param {Event} e
+     * @param {Object} building
+     */
+    handleMouseOver(e, building) {
+        clearTimeout(handleMouseOutTimeout);
+
+        if (mouseOverBuilding == building.id) return;
+
+        handleMouseOverTimeout = setTimeout(() => {
+            mouseOverBuilding = building.id;
+            super.handleEvent({
+                action: 'set-hovered-on-map',
+                payload: { buildingId: building.id }
+            });
+            if (building.selectOnMap !== true) {
+                super.handleEvent({
+                    action: 'set-selected-on-map',
+                    payload: { buildingId: null }
+                });
+            }
+        }, 200);
+
+        e.preventDefault();
+    }
+
+    /**
+     * Handles mouse leaves entry
+     * @param {Event} e
+     * @param {Object} building
+     */
+    handleMouseOut(e, building) {
+        const self = this;
+        clearTimeout(handleMouseOverTimeout);
+
+        handleMouseOutTimeout = setTimeout(() => {
+            mouseOverBuilding = null;
+            super.handleEvent({
+                action: 'set-hovered-on-map',
+                payload: { buildingId: null }
+            });
+        }, 100);
+
+        e.preventDefault();
     }
 
     /**
@@ -121,20 +177,31 @@ class Entry extends React.Component {
     }
 
     /**
+     * Toggle details of an a11y icon
+     */
+    a11yIconToggleShowDetails(e, name) {
+        const a11yIconShowDetails = this.state.a11yIconShowDetails;
+        a11yIconShowDetails[name] = ! a11yIconShowDetails[name];
+        this.setState({ a11yIconShowDetails });
+        e.preventDefault();
+    }
+
+    /**
      * Render entry, with title, icons and may details (address, note, ...)
      */
     render() {
         const building = this.state.building;
+        const a11yIcons = this.state.a11yIcons;
         const entryClass = building.selectOnMap ? "entry current-entry" : "entry";
 
-        // get all a11y icons
-        const a11yIcons = new A11yIcon({building: building});
-
         return (
-            <div id={`result-entry-${building.id}`} className={entryClass} onMouseOver={this.handleMouseOver} lang="de">
+            <div id={`result-entry-${building.id}`} className={entryClass} lang="de"
+                onMouseOver={e => {this.handleMouseOver(e, building)}}
+                onMouseOut={e => {this.handleMouseOut(e, building)}}
+            >
                 <h3><a href="#" onClick={e => {this.handleShowOnMap(e, building); this.handleShowDetails(e, building.id, false);}}>{building.title}</a></h3>
                 {!building.showDetails &&
-                    <ul className="a11yIcons-list">
+                    <ul className="a11yIcons-compact">
                         {a11yIcons.getAll().map((entry, id) => {
                             if (a11yIcons.icon(entry) == null) {
                                 return (null);
@@ -147,12 +214,26 @@ class Entry extends React.Component {
                 }
                 {building.showDetails &&
                     <div>
-                        <div className="a11yIcons-detailed">
+                        <div className="a11yIcons-extended">
                             {a11yIcons.getAll().map((entry, id) => {
                                 return (
                                     <Row key={id}>
                                         <Col xs={3} aria-hidden={true}>{a11yIcons.icon(entry)}</Col>
-                                        <Col xs={9} className="a11yIcons-descr">{a11yIcons.descr(entry)}</Col>
+                                        <Col xs={9} className="a11yIcons-descr">
+                                            {a11yIcons.descr(entry)}
+                                            {(a11yIcons.details(entry) !== null && this.state.a11yIconShowDetails[entry] === false) &&
+                                                <Button className="btn-link btn-toogle-a11y-details" aria-label="Deails" aria-expanded={false} onClick={e => this.a11yIconToggleShowDetails(e, entry)}>
+                                                    <i className="fa fa-chevron-down" aria-hidden={true}></i> Details
+                                                </Button>
+                                            }
+                                            {a11yIcons.details(entry) !== null &&
+                                                <Col className="a11yIcons-details">
+                                                    {this.state.a11yIconShowDetails[entry] === true &&
+                                                        <div>{a11yIcons.details(entry)}</div>
+                                                    }
+                                                </Col>
+                                            }
+                                        </Col>
                                     </Row>
                                 );
                             })}
