@@ -1,6 +1,8 @@
 import Promise from 'promise-polyfill';
-import JsonLoader from '../utils/JsonLoader';
+import HttpRequest from '../utils/HttpRequest';
 import {getDistance} from '../utils/GeoUtils';
+import {getGraphUri, getPrefix} from '../utils/RDF';
+import LocalCache from '../utils/LocalCache';
 
 /**
  * Stores all buildings
@@ -20,86 +22,268 @@ class BuildingStore {
         // default properties for each building
         this.buildingDefaults = {
             id: null,
+            title: '',
+            latitude: 0,
+            longitude: 0,
+            category:  '',
+            "entrance-suit-f-wheelchair": 0,
+            "lift-suit-f-wheelchair": 0,
+            "lift-avail": 0,
+            "toilet-avail": 0,
+            "toilet-suit-f-wheelchair": 0,
+            "parking-avail": 0,
+            "parking-f-disabled-avail": 0,
+            "help-for-hearing-imp": 0,
+            "help-for-blind": 0,
+            "general-help": 0,
+
             // is visible after apply filters
             visible: true,
             // is in bounds (northEast and southWest) of map
             inBounds: true,
             // show details of building
             showDetails: false,
-            // accessibility rating
+            // accessibility rating, the better the accessibility, the higher the ranking
             a11yRating: 0.00,
+            // ranking on search by title, 0: does not match, 1: partial match, 2: match
+            searchRank: 0,
             // use selected marker on map
             selectOnMap: false,
+            // use hover marker on map
+            hoveredOnMap: false,
             // object for loading data
             data: {}
         }
-        this.jsonLoader = new JsonLoader();
+        // http request interface to the triple store
+        this.httpRequest = new HttpRequest();
+        // local cache class
+        this.localCache = new LocalCache();
     }
 
     /**
-     * Init all buildings data
+     * Init all buildings with id, title, category, lift-avail ...
+     * Sort by accessibility
      *
      * @return Promise, length of added buildings
-     * @throws Error on loadJson
+     * @throws Error
      */
     initAll() {
-        // private function to add buildings to this.buildings
-        const addAll = (buildings) => {
-            for (const idx in buildings) {
-                const building = Object.assign({}, this.buildingDefaults, buildings[idx], {
-                    id: idx
-                });
-                building.a11yRating = this.getA11yRating(building);
-                this.buildings.push(building);
+        const self = this;
+        // sparql query to gel all buildings and some properties
+        const query = `SELECT ?id ?category ?title ?latitude ?longitude
+            ?entranceSuitableForWheelchairs
+            ?elevatorDoorWidth ?elevatorSuitableForWheelchairs
+            ?toiletDoorWidth ?toiletSuitableForWheelchairs
+            ?parkingLotOnSiteAvailable
+            ?numberOfSlotsDisabledParkingLotInFrontOfPlace
+            ?numberOfSlotsDisabledParkingLotOnSite
+            ?generalHelp
+            ?supportForHearingImpairedAvailable
+            ?supportForVisuallyImpairedAvailable
+        FROM <${getGraphUri()}>
+        WHERE {
+            ?uri <http://www.w3.org/1999/02/22-rdf-syntax-ns#type> <${getPrefix('place')}Place> ;
+                # match only public availables
+                <${getPrefix('place')}publicly-available> "true" ;
+                # properties
+                <http://purl.org/dc/terms/identifier> ?id ;
+                <http://dbpedia.org/ontology/category> ?category ;
+                <http://purl.org/dc/elements/1.1/title> ?title ;
+                <http://www.w3.org/2003/01/geo/lat> ?latitude ;
+                <http://www.w3.org/2003/01/geo/long> ?longitude ;
+                # parking
+                <${getPrefix('placeacess')}numberOfSlotsDisabledParkingLotInFrontOfPlace> ?numberOfSlotsDisabledParkingLotInFrontOfPlace ;
+                <${getPrefix('placeacess')}numberOfSlotsDisabledParkingLotOnSite> ?numberOfSlotsDisabledParkingLotOnSite ;
+                <${getPrefix('place')}parkingLotOnSiteAvailable> ?parkingLotOnSiteAvailable ;
+                # entrance
+                <${getPrefix('placeacess')}entranceSuitableForWheelchairs> ?entranceSuitableForWheelchairs ;
+                # lift
+                <${getPrefix('placeacess')}elevatorSuitableForWheelchairs> ?elevatorSuitableForWheelchairs ;
+                <http://semweb.mmlab.be/ns/wa#elevatorDoorWidth> ?elevatorDoorWidth ;
+                # toilet
+                <http://semweb.mmlab.be/ns/wa#toiletDoorWidth> ?toiletDoorWidth ;
+                <${getPrefix('placeacess')}toiletSuitableForWheelchairs> ?toiletSuitableForWheelchairs ;
+                # etc help
+                <${getPrefix('placeacess')}localSupport> ?localSupport ;
+                <${getPrefix('placeacess')}supportForHearingImpairedAvailable> ?supportForHearingImpairedAvailable ;
+                <${getPrefix('placeacess')}supportForVisuallyImpairedAvailable> ?supportForVisuallyImpairedAvailable .
+            # bind local support to true/false reduces data
+            BIND (
+                IF((?localSupport = ""), "false", "true")
+                as ?generalHelp
+            )
+        }`;
+
+        // get acc-value from sparql property
+        const getAcc = (str) => {
+            if (str === "fully") {
+                return 2;
             }
-            this.buildings.sort((a, b) => {
-                if (a.a11yRating > b.a11yRating) return -1;
-                if (a.a11yRating < b.a11yRating) return 1;
+            else if (str === "partly" || str === "true") {
+                return 1;
+            }
+            else {
                 return 0;
-            });
-            return this.buildings.length;
+            }
         };
 
-        return this.jsonLoader.loadJson('/building-coordinates.json')
-            .then((buildings) => {
-                this.initiated = true;
-                const added = addAll(buildings);
+        // private fct to add all buildings to this.buildings
+        const addAll = (buildings) => {
+            buildings.forEach((building) => {
+                const tmp = Object.assign({}, this.buildingDefaults, {
+                    id:                             building.id.value,
+                    title:                          building.title.value,
+                    latitude:                       parseFloat(building.latitude.value),
+                    longitude:                      parseFloat(building.longitude.value),
+                    category:                       building.category.value,
+                    "entrance-suit-f-wheelchair":   getAcc(building.entranceSuitableForWheelchairs.value),
+                    "lift-suit-f-wheelchair":       getAcc(building.elevatorSuitableForWheelchairs.value),
+                    "lift-avail":                   (parseInt(building.elevatorDoorWidth.value) > 0 || getAcc(building.elevatorSuitableForWheelchairs.value) > 0) ? 1 : 0,
+                    "toilet-avail":                 (parseInt(building.toiletDoorWidth.value) > 0 || getAcc(building.toiletSuitableForWheelchairs.value) > 0) ? 1 : 0,
+                    "toilet-suit-f-wheelchair":     getAcc(building.toiletSuitableForWheelchairs.value),
+                    "parking-avail":                (getAcc(building.parkingLotOnSiteAvailable.value) > 0 ||
+                                                     parseInt(building.numberOfSlotsDisabledParkingLotInFrontOfPlace.value) > 0 ||
+                                                     parseInt(building.numberOfSlotsDisabledParkingLotOnSite.value) > 0
+                                                    ) ? 1 : 0,
+                    "parking-f-disabled-avail":     (parseInt(building.numberOfSlotsDisabledParkingLotInFrontOfPlace.value) > 0 ||
+                                                     parseInt(building.numberOfSlotsDisabledParkingLotOnSite.value) > 0
+                                                    ) ? 1 : 0,
+                    "help-for-hearing-imp":         getAcc(building.supportForHearingImpairedAvailable.value),
+                    "help-for-blind":               getAcc(building.supportForVisuallyImpairedAvailable.value),
+                    "general-help":                 getAcc(building.generalHelp.value)
+                });
+                // get accessibility rating
+                tmp.a11yRating = self.getA11yRating(tmp);
+                self.buildings.push(tmp);
+            });
+            // sort buildings by accessibility rating
+            self.sortBuildingsBy('a11yRating');
+            return self.buildings.length;
+        };
+
+        // may get from cache (localStore)
+        const cachedBuildings = this.localCache.get("buildings");
+        if (cachedBuildings !== null && cachedBuildings.length > 0) {
+            const added = addAll(cachedBuildings);
+            if (added == 0) {
+                throw new Error('No buildings initiated');
+            }
+            self.initiated = true;
+            return new Promise((resolve) => {resolve(cachedBuildings.length)});
+        }
+
+        // new sparql http request to triple store
+        return this.httpRequest.request(query)
+            .then(function (response) {
+                const added = addAll(response.data.results.bindings);
                 if (added == 0) {
                     throw new Error('No buildings initiated');
                 }
+                self.initiated = true;
+                // add to local cache
+                self.localCache.set("buildings", response.data.results.bindings);
                 return added;
-            }).catch((ex) => {
-                throw new Error(ex);
+            })
+            .catch(function (error) {
+                throw new Error(error);
             });
     }
 
     /**
-     * Load data of a building and save to buildings
+     * Load data of a building and save to building.data
      *
      * @param {String} Id of building
      * @return Promise, true if buildings-data loaded
-     * @throws Error if loadJson fails
+     * @throws Error
      */
     loadBuildingData(id) {
-        // private function to save the data to building
-        const load = (data) => {
+        const self = this;
+        // sparql query to get a building and all its properties
+        const query = `SELECT ?p ?o
+        FROM <${getGraphUri()}>
+        WHERE {
+            ?uri <http://purl.org/dc/terms/identifier> "${id}" ;
+            ?p ?o
+        }`;
+
+        // fct to load properties to the building.data
+        const load = (buildingData) => {
             const building = this.getBuilding(id);
             if (building === undefined) {
                 return false;
             }
-            building.data = data;
-            return true;
-        };
+            building.data = {};
+            buildingData.forEach((entry) => {
+                switch (entry.p.value) {
+                    case 'http://purl.org/dc/terms/identifier':
+                        building.data.id = entry.o.value;
+                        break;
+                    case 'http://purl.org/dc/elements/1.1/title':
+                        building.data.titel = entry.o.value;
+                        break;
+                    case 'http://dbpedia.org/ontology/category':
+                        building.data.kategorie = entry.o.value;
+                        break;
+                    case getPrefix('geo') + 'lat':
+                        building.data.latitude = parseFloat(entry.o.value);
+                        break;
+                    case getPrefix('geo') + 'long':
+                        building.data.longitude = parseFloat(entry.o.value);
+                        break;
+                    case 'http://schema.org/addressLocality':
+                        building.data.ort = entry.o.value;
+                        break;
+                    case 'http://schema.org/streetAddress':
+                        building.data.strasse = entry.o.value;
+                        break;
+                    case 'http://schema.org/postalCode':
+                        building.data.plz = entry.o.value;
+                        break;
+                    case getPrefix('place') + 'openingHours':
+                        building.data.oeffnungszeiten = entry.o.value;
+                        break;
 
-        return this.jsonLoader.loadJson(`/buildings/${id}.json`)
-            .then((data) => {
-                const loaded = load(data);
+                    case getPrefix('placeacess') + 'localSupport':
+                        building.data.beschreibung_hilfestellungen_vor_ort = entry.o.value;
+                        break;
+                    case getPrefix('placeacess') + 'supportForHearingImpairedAvailable':
+                        building.data.besondere_hilfestellungen_f_menschen_m_hoerbehinderung_vorhanden = entry.o.value;
+                        break;
+                    case getPrefix('placeacess') + 'supportForVisuallyImpairedAvailable':
+                        building.data.besondere_hilfestellungen_f_menschen_m_seebhind_blinde_vorhanden = entry.o.value;
+                        break;
+                    default:
+                        // save all other properties with the hash of the uri
+                        const keyUrl = new URL(entry.p.value);
+                        building.data[keyUrl.hash.substr(1)] = entry.o.value;
+                }
+            });
+            return true;
+        }
+
+        // may get from cache (localStore)
+        const cachedBuilding = this.localCache.get(id);
+        if (cachedBuilding !== null) {
+            const loaded = load(cachedBuilding);
+            if (loaded == false) {
+                throw new Error(`Couldnt find building data with id "${id}"`);
+            }
+            return new Promise((resolve) => {resolve(loaded)});
+        }
+
+        // new sparql http request to the triple store
+        return this.httpRequest.request(query)
+            .then(function (response) {
+                const loaded = load(response.data.results.bindings);
                 if (loaded == false) {
                     throw new Error(`Couldnt find building data with id "${id}"`);
                 }
+                // add to local cache
+                self.localCache.set(id, response.data.results.bindings);
                 return loaded;
-            }).catch((ex) => {
-                throw new Error(ex);
+            })
+            .catch(function (error) {
+                throw new Error(error);
             });
     }
 
@@ -170,18 +354,6 @@ class BuildingStore {
     }
 
     /**
-     * Get specific building with given id
-     *
-     * @param {String} Id of building
-     * @return {Object} Building or undefined if not found
-     */
-    getBuilding(id) {
-        return this.buildings.find((building) => {
-            return building.id == id;
-        });
-    }
-
-    /**
      * Get all buildings
      *
      * @return {Array}
@@ -191,13 +363,34 @@ class BuildingStore {
     }
 
     /**
+     * Get specific building with given id
+     *
+     * @param {String} Id of building
+     * @return {Object} Building or undefined if not found
+     */
+    getBuilding(id) {
+        return this.getAll().find((building) => {
+            return building.id == id;
+        });
+    }
+
+    /**
      * Get all visible buildings
      *
      * @return {Array}
      */
     getVisibles() {
-        return this.buildings.filter((building) => {
+        return this.getAll().filter((building) => {
             return building.visible == true && building.inBounds == true;
+        });
+    }
+
+    /**
+     * Get selected building on map
+     */
+    getSelected() {
+        return this.getAll().find(building => {
+            return building.selectOnMap;
         });
     }
 
@@ -241,15 +434,24 @@ class BuildingStore {
      * @param {Array} Array with all filters
      */
     applyFilters(filters) {
+        let searchFilter = null;
         this.getAll().forEach((building, id) => {
+            let searchRank = 0;
             let isVisible = true;
             filters.forEach((filter, fid) => {
 
                 if (filter.type == 'search') {
-                    const r = new RegExp(filter.value, "i");
-                    if (building.title.match(r) === null) {
+                    searchFilter = filter;
+                    const part = new RegExp(filter.value, "i");
+                    const full = new RegExp("\\b" + filter.value + "\\b", "i");
+                    if (building.title.match(part) === null) {
                         isVisible = false;
                         return;
+                    } else {
+                        searchRank = building.title.match(full) === null ? 1 : 2;
+                        if (filter.value == "") {
+                            searchRank = 0;
+                        }
                     }
                 }
                 else if (filter.type == 'select-one') {
@@ -267,8 +469,17 @@ class BuildingStore {
                 }
 
             });
+            building.searchRank = searchRank;
             building.visible = isVisible;
         });
+        if (searchFilter !== null && searchFilter.value !== "") {
+            // sort by searchRank and a11Rating
+            this.buildings.sort((a, b) => {
+                return b.searchRank - a.searchRank || b.a11yRating - a.a11yRating;
+            });
+        } else {
+            this.sortBuildingsBy('a11yRating');
+        }
     }
 
     /**
@@ -328,11 +539,28 @@ class BuildingStore {
      */
     setSelectedOnMap(id) {
         this.getAll().forEach((building) => {
-            let isSelected = false;
-            if (building.id == id) {
-                isSelected = true;
-            }
-            building.selectOnMap = isSelected;
+            building.selectOnMap = building.id == id;
+        });
+    }
+
+    /**
+     * Sets flag for a building to use selected marker on map, only one building will be selected
+     *
+     * @param {String} Building id
+     */
+    setHoveredOnMap(id) {
+        this.getAll().forEach((building) => {
+            building.hoveredOnMap = building.id == id;
+        });
+    }
+
+    /**
+     * Sort builidings numerical DESC by a given key
+     * @param {String} key
+     */
+    sortBuildingsBy(key) {
+        this.buildings.sort((a, b) => {
+            return b[key] - a[key];
         });
     }
 }
