@@ -1,10 +1,15 @@
 import React from 'react';
 import sinon from 'sinon';
 import Promise from 'promise-polyfill';
-import {shallow, mount} from 'enzyme';
+import {configure, shallow, mount} from 'enzyme';
+import Adapter from 'enzyme-adapter-react-16';
+configure({ adapter: new Adapter() });
 
-import {_10buildings} from './assets/10buildings';
-import {fetch} from './utils/fetch';
+//import {_10buildings} from './assets/10buildings';
+//import {fetch} from './utils/fetch';
+import mockedAxios from 'axios';
+import _10buildings from './assets/10-buildungs-index'
+
 import {getLogger, getStores, getEventHandler} from './utils/utils';
 import {wrapBuildingNavigator} from './utils/wrapComponents';
 
@@ -12,33 +17,30 @@ import BuildingNavigator from '../src/BuildingNavigator';
 import Map from '../src/components/Map';
 import Marker from '../src/components/map/Marker';
 
+jest.useFakeTimers();
+
 describe('<Map />', () => {
 
-    // we cannot mount the map as root node! (see: https://github.com/PaulLeCam/react-leaflet/issues/246)
+    it('renders markers', async () => {
 
-    it('renders markers', () => {
-        const buildingNavigator = wrapBuildingNavigator('mount');
+        const logger = getLogger();
+        const stores = getStores(logger);
+        const eventHandler = getEventHandler(stores, logger);
 
-        const eventHandler = buildingNavigator.instance().eventHandler;
+        const map = mount(
+            <Map stores={stores} logger={logger} eventHandler={eventHandler} />
+        );
 
-        // create promise for initall()
-        const promise = buildingNavigator.state().stores.buildingStore.initAll();
+        mockedAxios.mockImplementationOnce(() => {
+            return Promise.resolve({ data: _10buildings });
+        })
 
-        // fake handleEvent()
-        sinon.stub(eventHandler, 'handleEvent').callsFake((event) => {
-            if (event.action === 'init-buildings') {
-                return promise;
-            } else {
-                return Promise.resolve();
-            }
-        });
+        const promise = await stores.buildingStore.initAll();
 
-        // initAll() is an async method, so we test after promise resolves
-        return promise.then(() => {
-            buildingNavigator.update();
-            // currently all buildings are on the same location -> it one grouped marker
-            expect(buildingNavigator.find(Marker).length).toBe(1);
-        });
-    })
+        expect(map.find(Marker).length).toBe(0);
+        map.setProps({stores});
+        expect(map.find(Marker).length).toBe(10);
+    });
+
 
 });
